Driver core: Add option to suppress automatic binding
Some time ago there were a couple of discussions on the Linux kernel
embedded mailing list about techniques to suppress the binding of
device drivers to devices until a point the user chooses ([1] and [2]).
In general the conclusion of both discussions was that something in
exported via /sys might be the way to go.

We also had a requirement for something similar in a project here. In
our case it was for a system-on-chip device, where some peripherals
could be managed by either the processor running Linux or from another
processor on the device, and we wanted a technique which didn't
require recompiling the kernel to make this switch.

To do this I've added a new field (suppress_bind) to the device
structure. When set this prevents a driver being bound to the device.
This field can either by set directly (for example, when adding a
platform device) or through a command line argument, for example:

    suppress_bind=stm-ohci.0,stmfdma

If at a later time the user wants to bind a driver to the device,
this flag can be cleared:

    echo 0 > /sys/devices/platform/stm-usb.0/stm-ohci.0/suppress_bind

and a normal dynamic bind triggered:

    echo -n  stm-ohci.0 > /sys/bus/platform/drivers/stm-ohci/bind

[1] http://article.gmane.org/gmane.linux.kernel.embedded/578
[2] http://article.gmane.org/gmane.linux.kernel.embedded/389

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>