sh: Only flush code in signal frame if it was successfully written
This fixes a problem seen when sending a signal to a process which has
registered a signal handler and the signal hander return code is poked
onto the user's stack, but the stack pointer is invalid. In this case
the code which writes the code onto the stack is safe, as it uses
__put_user(), but the code to flush the I cache isn't as it assumes
the address passed in is valid.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>