mmc: update the whole framework to the version from 2.6.37.6
This patch updates the whole MMC stack to the version
included into the kernel 2.6.37.6.
This has been possible because a sub-set of mmc patches
were already added in this kernel 2.6.32.
Too many patches could be added by using git cherry-pick
but this step wanted effort and too many conflict had to
be manually solved. The best strategy for me was to directly
merge the new stack for the mmc that includes many new
features and will make easier the commit of new patches
from the mmc kernel development.
In the end, this patch actually is self-contained and
it cannot generate problems on other parts.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>