sh_stm: Fix PCI support for hdk7111 board
Update the PCI support for the hdk7111 board. There were a number of
problems with this including not using the IO machvec and not working
around the fact that idsel is driven by a PIO rather than an address
line. The board  also has a very strange irq layout which connects what
is usually INTA to INTC, so it requires a custom irq routine.

This patch also corrects the serial port support, the serial port does
not have hardware flow control, and ASC3 is not connected up to a serial
port so is removed.

This board was also not included in the pci rework, so was including
pci-synopsys.h which no longer exists. This unfortunately caused
compilation errors even if you were not using PCI.

Signed-off-by: David McKay <david.mckay@st.com>