mmc: core: eMMC 4.5 Power Class Selection Feature [v5]
This patch ports the following recent patch from the mmc dev.

This patch adds the power class selection feature available
for mmc versions 4.0 and above.
During the enumeration stage before switching to the lower
data bus, check if the power class is supported for the
current bus width. If the power class is available then
switch to the power class and use the higher data bus. If
power class is not supported then switch to the lower data
bus in a worst case.

Signed-off-by: Girish K S <girish.shivananjappa@linaro.org>
Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>