sdhci: pltfm: add get_max_clk hook and clock support
This patch adds the get_max_clk hook to pass from the
platform the real clock used for the host controller.
By default, Arasan suggests to use 48MHz; on STM platforms
the HC is fully validated at 50MHz that is no std clock.
This has to be passed to the main sdhci driver and so the
new hook has been added into the platf sdhci d.d.
Also this patch adds the clock support to disable/enalbe
them on suspend/resume functions.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>
Reviewed-by: Francesco Virlinzi <francesco.virlinzi@st.com>