stm_clk: introduce a new alias for MMC on 7108,7105, 5206
The SDHCI driver is now able to use no std clock but a
new clk alias is required to well detect the hc clock
into the mmc sdhci platform d.d.
Note that, it's unclear the reason of two clocks in case
of the 7108/05. In fact, card clock should be always provided
by the HC.
Arasan suggests to use 48MHz but we validate our HW with 50MHz
that is ok because < 52MHz.
Depending on the card actually used, the HC clock has to be scaled
(by using own HW register and calculating some divisors). So the HC
is able to provide the good clock to the card.
The patch takes only care of the clock of 50MHz that will be
passed to the mmc layer.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>
Reviewed-by: Francesco Virlinzi <francesco.virlinzi@st.com>