stm_stx5206: fix mmc clk
After adding the patches to propagate and manage
the right clock to the mmc device driver, I noticed
a problem on the hdk5206 where the default clock
provided by PLL is not 50MHz but 27MHz.

In the past, the HC didn't verify the clock and used
the std clock set to 48MHz (always supposed good!).
This could generate problems when handling card in
High speed.
This patch verifies the clock from the PLL and, in
case of broken value, fix it to 50MHz.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>
Reviewed-by: Francesco Virlinzi <francesco.virlinzi@st.com>