stm_mali: Allow Mali access to all of both LMIs
The Mali "EXTERNAL_MEMORY_RANGE" resource is used to validate addresses
passed into the driver. However, it is difficult to construct a simple
address range which can cope with differing amounts of memory in both
LMI0 and LMI1 and the options for LMI0 to be populated from the bottom
up (as is conventional) and from the top down (so it is contiguous
with LMI1). In most general the kernel doesn't know or care about these
differences.

So to keep things simple, this change makes the whole of LMI0 and LMI1
valid. There is some loss of error checking by doing this, but common
errors such as 0 and 0xffffffff will be caught.

See bug 12369 for further discussion of this issue.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>