sh_stm: Add EMI 'Write-Protect' mapping
This patch adds support for a 'Write-Protect' mapping of an EMI Bank.  The
mapping, which is intended to be used in conjunction with the Physmap driver,
disables the EMI 'Write-CS' when not required by the driver.  This provides some
level of protection against rogue accesses to the EMI bank (see bugzilla #15001
for one such example).

To use this feature, enable the configuration options MTD_COMPLEX_MAPPINGS and
MTD_STM_EMI_WRITPRO_MAP, and update the phymap platform data to specify the map
initialisation function, for example:

	board_physmap_flash_data.map_init = stm_emi_map_init,

Signed-off-by: Angus Clark <angus.clark@st.com>
Signed-off-by: Chris Smith <chris.smith@st.com>