stm: stxh205: Add board specific config for PHY daughter boards
Both the B2039 and B2057 currently have the Ethernet PHY attched via
a daughter board. A couple of different boards are available, which
require different Ethernet configuration parameters, so add Kconfig
options to allow the board to be selected.

In addition for the B2039 add an option for the Ethernet PHY reset,
as this may not be connected depending on jumper settings, and as it
conflicts with other usage of this pin it may be useful to remove
the software reset and rely on the board level reset signal alone.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>