b2057: Add support for STxH207 with internal IC+101A PHY
Some versions of the STxH207 now have an internal IC+101A Ethernet
PHY. Add support for this to the B2057.

Note that unlike the B2039 this doesn't support the 'no external clock'
configuration, because its not clear what happens to RX_CLK/C50M_O
pin, and as there is no way to disconnect it there may be a risk of
contention.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>