stm: pad: Add stm_pad_update_gpio()
Add an API to the stm pad manager which allows the configuration of
a pad to be changed at run time. Effectively this is just a wrapper
around the call to the SoC specific pad configuration function, which
allows the direction, function or retiming to be changed at run time.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>