kptrace: Add a relay control file to control buffer overwrite behavior
This patch adds a relay control file to debugfs, allowing userspace to
specify whether or not KPTrace's relay sub-buffers should be overwritten
if they are all full.

A patch follows which makes use of this to vary the behavior between
"normal" and "flight-recorder" modes.

Note: the struct file_operations is const because checkpatch, rightly,
says that it should be. Another patch follows to make the others const
too.

Signed-off-by: Chris Smith <chris.smith@st.com>