mm: Allow a default value for min_free_kbytes to be set using Kconfig
The default value of min_free_kbytes is a function of the size of system
memory, and it can then be changed using procfs. This patch allows a
different default value to be specified using Kconfig, which is useful
for users who require a different figure but do not have procfs mounted.

Signed-off-by: Chris Smith <chris.smith@st.com>