pci_stm: Move shared EMISS configuration from pci-emiss to emi
Previously, the pci-emiss driver grabbed a large chunk of the EMISS address
space.  This included shared configuration registers, and in particular, the
NAND-BCH DMA register space.  This patch moves configuration of the shared
registers to the emi driver, and restricts the pci-emiss driver to the
PCI-Bridge region.

    	emiss			offset		owner

    	emiss_empi		0x0000
    	emiss_nanddma		0x0800		[stm-nand-bch]
    	emiss_config		0x1000		emi
    	emiss_pcibridge		0x1400		pci-emiss

We also fix the "emi4 config" base address for stx5197 and stxh205.

Signed-off-by: Angus Clark <angus.clark@st.com>