mtd: nand: update READID decode
This patch updates the code that derives device properties from the READID data.
With the increasing number of encoding schemes and exceptions that must be
handled, the code was becoming difficult to read and maintain.  Here, we
refactor the code according to three basic schemes:

	'ID 2'       : extract properties from device table.  For 2-byte IDs, or
                       where the device table gives a non-zero page-size (SP
                       devices).

	'Extended ID': decode properties from ID string where possible, falling
	               back to device table.  For 3/4/5-byte IDs.

	'ID 6'       : decode properties from ID string where possible, falling
	               back to device table.  For 6-byte IDs.

The new code has been tested on a number of real NAND devices, and within a
standalone test harness using the data found in
http://www.linux-mtd.infradead.org/nand-data/nanddata.html.

To avoid further cluttering nand_base.c, and to allow other, non-standard, NAND
drivers to make use of the code, it is moved to nand_id.c.

(An equivalent patch is to be submitted to mtd-2.6, hopefully for inclusion in
linux 3.3.)

Signed-off-by: Angus Clark <angus.clark@st.com>