stm_nand: updates to NAND controller register definitions
In preparation for the 'stm_nand_bch' driver, this patches updates the NAND
controller register definitions in-line with the latest draft of the NANDi
functional spec ('nandi_bch_ctrl' v2.4).

The NANDi controller comprises the Hamming Controller (FLEX/AFM) and the new BCH
Controller (AFM-only).  Both controllers share the same register space with a
MUX to route access.  Although the offsets, names, and fields of many of the
registers are the same, there are one or two exceptions.  To avoid confusion,
two sets of registers are defined, one set for the Hamming Controller, and one
set for the BCH Controller.

The existing drivers supporting the Hamming Controller (stm_nand_flex and
stm_nand_afm) are updated to use the new register definitions.

Finally, we update the named IO resource "flex_mem" to "nand_mem", since it is
used by the stm_nand_flex and stm_nand_afm (and stm_nand_bch) drivers.

Signed-off-by: Angus Clark <angus.clark@st.com>