stm_nand_afm: update PBL boot-boundary code
Previously, the code that extracted the PBL "boot-boundary" assumed that the PBL
was located at the start of logical block zero.  However, updates to the PBL
mean this assumption is no longer valid.  Instead, we must search for the PBL
"Magic Markers" and Offset in order to determine the PBL location.  This patch
updates the boot-boundary code accordingly.  We also update the read/write
routines to handle data requests that straddle the boot-boundary.

Signed-off-by: Angus Clark <angus.clark@st.com>