stxh205: disable boot from eMMC to allow access to ARASAN
On the B2039 when we set the mode pin to boot from eMMC
then the Kernel, at run-time, will stuck as soon as it probes
the SDHCI driver because the ARASAN HC cannot be used at all.

So, to disable boot from eMMC and allow the access to ARASAN
MMC/SD host controller we need to reset the
MMC_BOOT_ENABLE_CONFIG_BIT in the MMC configuration register 2.

Reported-by: Youssef TRIKI <youssef.triki@st.com>
Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>