mmc: sdhci: do not drop probing step when clk_get fails.
On ST platforms the SDHCI drivers is able to get the clock
from the platform because a custom clk source can be passed to the HC
by using the SDHCI_QUIRK_CAP_CLOCK_BASE_BROKEN flag.
Without this setting, also suggested by the Validation, the HC will use
the default 48MHz for (HS - high speed) cards instead of 50MHz.
This approach cannot be used on boards like Freeman where the clocks are
provided from another output source (ckg_south) not yet supported.
Obviously, we suppose that the clock provided from the targetpack
is good also for HS cards (and this seems to be true after some first tests).
>From the Freeman Clock Generator Functional Specification mmc card clock
should be 52MHz but not actually measured on my side.

W/o this patch the new SDHCI driver will fail on Freeman:
see STLinux Bugzilla Bug 17143.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>
Reviewed-by: Francesco Virlinzi <francesco.virlinzi@st.com>