stm: Move IC+1001 delay configuration to a phy_fixup
On the STx7108 boards which use an IC+1001 Ethernet PHY, it is necessary
to configure the PHY with no additional delays on RxCLK and TxCLK when
using a GMII interface, and with additional delays on both when using
RGMII interface.

Previously we assumed that the pull/up down resistors on the board
were correctly set for GMII and the PHY driver used the MDIO bus to
set the delays for RGMII. The problem with this is that it didn't
allow for different boards which had different requirements, and it
also assumed that the boards using GMII were correct, which in the
case of the HDK7108 it appears it is not.

So move this code out of the PHY driver and use the phy_fixup
infrastructure to change this only on the boards which require it.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>