serial: stm_asc: Avoid spurious interrupts in console output
If normal serial port output is interrupted and the interrupt handler
performs a printk which is directed to a console on the same serial
port, there is a danger that it will interfere with the interrupt state
of the serial port.

So take a much more cautious approach when performing the output:
 - save the interrupt mask and disable all interrupts
 - perform the output
 - wait for the FIFO to drain
 - restore the interrupt mask
This may cause a previously asserted interrupt to be de-asserted, but
it will be restored when console output is completed, and more
importantly no interrupts should be spuriously asserted.

This code is now modelled after serial8250_console_write(). In particular
the locking is fixed so that we don't deadlock when using sysrq or if
a panic is occurring.

See bugzilla 15934 for one way to reproduce this problem.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>