stm: b2057: Configure power control pin
This patch configures Power Control pin PIO(3.7) on B2057 board in
output mode.

On B2057 schematics this pin is labelled 7108_HPS.

Setting this line low enables the SoC power (it is pulled low by
default). This can then be used to control power to the board during
controller passive standby.

Normally the host CPU is in control of this pin, so performs the PIO
configuration.  However when powering down the SBC takes control of
this pin, but the firmware expects the PIO configuration should have
been done by the host.

See bugzilla 15904.

Signed-off-by: Udit Kumar <udit-dlh.kumar@st.com>