stm:fli7560: Add support for Freeman Premier 2
This patch adds support for the Freeman Premier 2 (aka FLI7560).

The FP2 closely resembles the Freeman Lite (aka FLI7510) in many ways,
but to avoid too many overcomplicated if-elses, we just switch on the
cpu type, allowing FLI7510/60 and 20/30/40 (Freeman Ultra variants) to
be easily grouped together. This will help if/when we support the
FLI7515 and FLI7525.

Note that the FLI7560 can only be differentiated from the FLI7510 by
its devid, which is only available once sysconf is configured, so we
have to initially probe it as an FLI7510 and then check the devid in
fli75xx_early_device_init().

Note also that this patch is based on my earlier submitted work
replacing fli7510_* with fli75xx_* where appropriate.

Signed-off-by: Chris Smith <chris.smith@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>