usb: hcd: stm: fix system crash when unloading hcd_stm.ko
This patch solves a system crash occurring while trying to unload
the hcd_stm.ko with the {e,o}hci_hcd.ko still loaded.

When the hcd_stm.ko is unloaded the function 'stm_usb_remove' is called.

It does few actions:
 - power down USB ports
 - disable clocks
 - unregister the nested ehci/ohci devices

When the ehci/ohci drivers are loaded, through the device model
the related {ehci,ohci}_hcd_stm_remove callback is called, as described
by the call stack below (let's refer to the ehci example)

stm_usb_remove()
\_ stm_device_power()     <-- host and phy powered-off
|_ stm_usb_clk_disable()
|_ platform_device_unregister(ehci_device)
   \
     ...
        \_ device_release_driver()
           \_ ehci_hcd_stm_remove()
              \_ usb_remove_hcd()
                 \_ hcd->driver->stop() [==> ehci_stop()]
                    \_ ehci_quiesce() <-- attempts to access USB registers

Here any attempts to read the hardware regsisters by
ehci_readl(ehci, &ehci->regs->command) caused a system crash, because the
hardware has been already powerd-off so it is not longer accessible.

When the ehci_hcd.ko (or ohci_hcd.ko) is not loaded, the code path starting
from device_release_driver() is not taken because there is not driver bound
to the platform device, so everythings is fine.

If the ehci_hcd.ko (or ohci_hcd.ko) is unloaded prior to hcd_stm.ko, USB is still
powered, so this issue is not spotted.

So the fix consists in moving calling the platform_device_unregister() before
stm_device_power().

Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>