stm: hom: gpio: Leave the gpio_irq enabled in case of HoM
Normally it would not be necessary to enable or disable the gpio
pins when entering HoM as the SoC is going to have power removed and
this state would be lost.

The problem is that the system has to guarantee the gpio_irq is
enabled for the SBC (not for the CPU)

This is what this patch does.

Reference: Bugzillla 18162

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>