stm: fdma: improve FDMA reset procedure
This patch revises the FDMA reset procedure by folding the two functions
fdma_disable_all_channels() and fdma_reset_channels() into the new
fdma_reset_all() function.

Further this has been extended to reset the CPU program counter and all the
instruction pipeline registers and keep the clock gated so that no further
instructions are executed.

This fix is needed to be able to boot a new kernel (with FDMA configured)
by kexec.

Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>
Signed-off-by: Angelo Castello <angelo.castello@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>