stm: fli75xx: Fix pio config for UART3 TX
All the RT/TX pins for the 3 uarts on Freeman 7510 are alternate
functions according to the data sheet.

However the pio config for the TX pin of UART2 (unlike UART1 and UART2)
sets STM_GPIO_DIRECTION_OUT rather than STM_GPIO_DIRECTION_ALT_OUT

As a result - UART3 is non-functional

   See Bugzilla 20175

This patch changes the config of the relevant pin to the alternative function.

Signed-off-by: Jon Frosdick <jon.frosdick@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>