stm: coprocessor: Fix validating BSS section of ELF files
The BSS section of ELF files appears as a normal section, but it is not
stored in the ELF file to save space, and is filled with zeros when the
file is processed. As a result there is no point in validating that
the contents are within the file, as they won't be.

This is a fix for Bugzilla #19404:
"cannot load SDK2 audio or video firmware"

Signed-off-by: Adrian Thomasset <adrian.thomasset@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>