stm_stxh205: fix pm in Lille by enabling the eth clk
This patch enables the eth.clk and the eth_phy.clk to guarantee
the parent of these clock is not erronously disabled while suspending.
Without this, a problem is seen while suspending when walking the
device tree. When the CPU turns-off the MMC-clock the parent PLLs has no
enabled children and so the PLL is turned off, but the the CPU
hasn't yet suspended the Ethernet.

This patch just declare the eth-clk is on.

Without this patch the kernel stucks when suspending.

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>
Tested-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>