stm: b2076: Initial B2076 board support
Support for the B2076 STxH207-HVK processor module. This is a minor
variant of the B2039, the two major changes being:

 - support for the STxH207xKx (LiegeC), which is a different ballout
   using an FPBGA instead of the previously supported STxH207xUx
   (Liege) which uses a PBGA package. This new ballout is compatible
   with the Lille MCM variants with integrated frontend, such as
   STxH237 (Cardiff), STxH238 (Newport) and STxH239 (Bristol).

 - a PCIe slot instead of a SATA connector. However using a small
   converter board this can be used for SATA as well.

Note that although the layout of this board is very similar to the B2039,
many components have different numbers, which means that connectors,
jumpers and other component numbers may have changed.

It is assumed that all Lille variants designed for this board will have
in in-package IC+101G PHY so support for an external PHY has not been
provided.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>