stm_b2076: Add PCIe support for B2076
Enables support for the PCIe slot on the B2076 HVKA board. This slot can
also be used for a SATA connector using the special adaptor board.  A
kconfig option selects which one is desired as they are mutually
exclusive and cannot be probed.

The PCIe reset signal #PERST is connected to PIO 4,6  J29 needs to be in
position 2-3 to use this signal. It is multiplexed with SSC1, so in most
cases that will be used instead. Fortunately PCIe cards must implement a
POR so the use of the #PERST signal is not mandatory.

Signed-off-by: David McKay <david.mckay@st.com>