stm_stxh205: Update USB AMBA plug settings
Update the USB AMBA wrapper configurations to the latest settings
from validation. We have to set it differently for 16 or 32 bit
DDR configurations, so we probe the LMI registers to find out which
mode we are running in.

There are some new USB flags introduced, but other SOCs should be
unaffected by the changes as the SOC code only sets the flags and
does not manipulate them.

Signed-off-by: David McKay <david.mckay@st.com>