stm: clk: Prevent unbalanced calls to clk_disable()
This patch adds a control in clk_disable() to guarantee the clk
isn't modified when the usage_counter is equal to zero.

This ensured that unbalanced enable/disable calls don't cause a problem.

Now the code is raising a WARN_ON in case of clk_disable on usage_counter
equal to zero.

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>