stm: pio-control: Handle zero values in retiming config correctly
Each field of the retiming config structure can take the value -1
to indicate it should not be changed, or the value to set it to,
which is typically 0 or 1. However, the code was previously treating
a value of 0 as don't care. This led to register values which should
have been changed from 1 to 0 being left at 0.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>