stm: h205: clk: CLK_B_REF supports enable/disable
This patch fixes an issue in the current clk-h205.
The CLK_B_REF clk (parent for all the clock-gen-B tree)
didn't support enable/disable (because marked as
always_enabled).

But moving on a system without always_enabled marker
the CLK_B_REF fails to enable and it's usage_counter
remains to zero.
For this reason the system is raising a warning during
the shutdown.

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>