sata_stm: Add dependency on MiPHY
As the STM SATA driver calls into the MiPHY code directly, it really
is dependant on it. Given that we now only enable STM_MIPHY on those
devices which actually have the hardware, it makes sense to add an
explicit dependency in the SATA_STM configuration on STM_MIPHY,
to avoid compilation errors if the SATA driver is accidentally enabled
in a build which doesn't have MiPHY enabled.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>