ide-cd: Fixup build breakage
Commit 2ae109f52ede318670914b21a038471d04708ca0
    cdrom: Fix ATAPI packet commands using DMA buffers on the stack
changed ide_cdrom_get_capabilities() so that the capabilities, which
were frequently on the stack, were not written directly using DMA,
but were instead first written to a kmalloced buffer before being
copied to the address passed in by the caller.

However, subsequent commit 455d80a9553915251b6392e31c39aeb8f8a94d68
    ide-cd: remove struct atapi_capabilities_page (take 2)
changed the name of this parameter.

When this was merged this was not spotted, so fix it now.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>