phy: add the WoL support and allow the access to paged registers
This patch adds a new feature called WoL+ and it is
able to manage a suspend mode present in some new
PHY devices (e.g. IC+101G).

To set this kind of feature some special registers
have to been read/write. So the patch adds new API
to access to these registers.

Also the patch adds the interface to manage the WoL
via ethtool passing from the Ethernet driver.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>