stmmac: using the WoL+ capability
If the stmmac understands that the PHY supports the new
WoL+ feature so it won't perform this job to wake up the system
at all and, during the suspend, its core and dma will be powered-off.

To make easier the usage of this new WoL+, the stmmac
will guarantee the ethtool interface.

This means that, at user-level, nothing is changing and
to set the WOL via magic frame this commands will be done:

  ethtool -s eth0 wol g
  echo standby > /sys/power/state

V2: added a new module parameter to declaire if the stmmac
wants to use the WoL+ feature provided by the PHY.
This allows us to spare any ifdef options.
Next step could be to have it in the ethtool

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>