phy: icplus: add WoL plus support
Some new ICPLUS PHYs like the 101G support the Wake-Up on Lan capability.
The IC+101G PHY can enter in a suspend mode after a period time
(default is 3min).
When the PHY enters in WoL+ mode the link speed is also
reduced to 10M. When the PHY is waked-up the original link
is restored.
Also this option allows the PHY to declare to the upper-layers
that it is able to wake-up the system for example via
magic packet. In this scenario the MAC can decide to demand
this to the PHY and the MAC core can be totally powered-off.

V2: as Stuart suggested: I removed the Koption to enable this
feature. This will be on by default. So the 101G can enter in
the new power-save mode. This should not effect the 101A.
Concerning the feature to wake-up the system the stmmac can
decide to use this PHY device.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>
Cc: Stuart Menefy <stuart.menefy@st.com>