stmmac: prevent null pointer when look at is there is pending work
This patch adds a safety check when the driver verifies is there is
pending work. In fact, in some circumstances (not easy to reproduce),
on customer's box, when the TIMER optimization is enabled, there
are spurious interrupts and the has_work function is called when
the DMA rings are not yet initialized. The patch so proposes a simple
work-around that verifies if dma_rx is not NULL before getting the OWN
bit on the RX current descriptor.
Note that the TIMER optimization is quite old and not used in new kernel
series. So IMO this is the best work-around to this strange phenomenon
(never seen on other ST platforms indeed).

Thanks go to Benjamin Colson for having tested it and provided me
his support on reproducing the issues.

See Bugzilla Bug: 21311

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>