stm: lpm: Driver for the Standby Controller
This patch adds a driver for the Low Power Monitor (LPM) to
communicate with Standby Controller (SBC).  For HDK7108 SBC is
external to SOC, whereas for STxH207 this is internal to the SOC.

Communication between SBC and host CPU is done using messages.

Each message has a common part, and a command specific part.

The common part contains the following information:
a) A command identifier.
b) A transaction identifier. This is a incrementing number that is
   used to pair up commands and responses.  This is important since
   there can be multiple host CPU's talking to the SBC CPU.

The specific part of a message depends on the command, and whether it
is a request or a response.

Details of communication protocols are in ADCS 8386373 A.

In case of internal SBC, communication will be done using a mailbox.
The mailbox has a size of 16 bytes, message greater than 16 bytes are
called big message, and will be sent using the SBC's DMEM.  Big
messages are most frequently seen when sending the IR data on which
SBC needs to wakeup.

In case of external SBC (eg 7108) communication is performed using
I2C.

Signed-off-by: Udit Kumar <udit-dlh.kumar@st.com>