stm_nand: make STM NAND drivers hot-pluggable
Although NAND devices are not considered to be hot-pluggable, there are cases
where it is useful to register/unregister the device at runtime.  Furthermore,
the driver model may make its own assumptions on when to call the initialisation
functions.

With this in mind, we update the driver initialisation functions from '__init'
to '__devinit'.  The '__exit' functions are updated accordingly.  The patch also
fixes a memory leak in the stm-nand-flex driver when removing as a module.

Signed-off-by: Angus Clark <angus.clark@st.com>