mtd: add flags to help identify MTD partitions that span the entire device
A MTD storage device may be registered with the MTD framework either as single
'master' device, or as one or more 'slave' partitions (which may not necessarily
cover the entire device).  Each device/partition registered is exported to
user-space through an /dev/mtd? device node.

Unfortunately, it is not possible to determine how each exported MTD device maps
onto the physical devices upon which it is hosted.  Ideally we would like to
link each MTD device to its master, and the offset within that master.  However,
this would require significant changes to the way in which devices are
registered, and to the information exported by MTD through sysfs.

For our immediate needs (related to the STM NAND Tools package) all we need to
know is whether the MTD device spans the entire physical device (thereby
allowing a one-to-one mapping between MTD offsets and absolute offsets on the
device).

To support this requirement, with minimal changes, we add two new flags,
MTD_SLAVE_PARTITION and MTD_SPANS_MASTER.  These can be accessed from user-space
via the existing MEMGETINFO IOCTL.

Signed-off-by: Angus Clark <angus.clark@st.com>