stm_nand_afm: no longer assume AFM is quiescent mode of operation
The STM NAND Hamming Controller supports two modes of operation, FLEX mode and
Advanced FLEX mode (AFM).  Whilst the stm_nand_afm driver is based predominantly
on AFM, it is sometimes necessary to fallback to FLEX mode for certain
operations.

Previously the driver assumed AFM was the quiescent mode of operation; AFM was
selected during initialisation, and we would always return to AFM following any
operation conducted in FLEX mode.

This patch removes any assumption over a quiescent mode of operation.  The
appropriate mode is selected prior to any operation, and no attempt is made to
return the controller to the previous state.

The patch makes no change in overall functionality, but it paves the way for
subsequent patches aimed at re-using more of nand_base.c, where it would become
rather complicated to ensure the controller is always returned to AFM.

Signed-off-by: Angus Clark <angus.clark@st.com>