stm_nand_bch: update handling of MTD OOB operations
The BCH NAND Controller was designed on the premise that nothing should use the
OOB area for storing meta-data.  Indeed, in many cases, the entire OOB area is
consumed by the ECC data and there is no "free" space available.  As such, the
BCH Controller provides no direct access to the OOB area, and the driver returns
"no free space available" to the clients in all cases.

In order to support some of the essential OOB operations, such as scanning for
manufacturer-programmed bad block markers, the stm-nand-bch driver falls back to
Hamming-FLEX mode.  For the most part, these OOB operations are restricted to
internal use by the driver.  However, we do attempt to provide some limited
support for the mtd_read_oob() and mtd_write_oob() callbacks.  This can be
useful for obtaining "raw" dumps of a page and performing various ECC tests
during development.

Previously, the mtd_read_oob() and mtd_write_oob() implementations would test
the OOB mode flag to determine if the requested operation made sense in the
context of the stm-nand-bch driver: 'MTD_OOB_RAW' is considered valid, for
returning the 'raw' Page and OOB data, whereas 'MTD_OOB_AUTO' and
'MTD_OOB_PLACE' are semantically ill-defined, given the interleaved nature of
data and ECC sectors when ECC is employed, and the fact that the driver reports
there is no "free" space available for meta-data.  In this case, the calls
return with '-ENOTSUPP'.

However, it turns out that some clients (e.g. YAFFS2 with 'inband-tags' enabled)
use the mtd_read_oob() and mtd_write_oob() calls, and even set the OOB mode
flag, for operations that do not actually involve OOB data at all.  (In the case
of YAFFS2, this is probably down to the way in which 'inband-tags' support was
tagged on to the existing code.)

This patch updates the mtd_read_oob() and mtd_write_oob() implementations, such
that 'MTD_OOB_PLACE' and 'MTD_OOB_AUTO' modes are only rejected if the requested
operation actually includes OOB data, as indicated by a non-NULL 'oobbuf'
pointer being passed in.  Furthermore, if 'oobbuf' is NULL, and the mode is not
'MTD_OOB_RAW', then the request is passed on the to more conventional callback,
mtd_read() or mtd_write(), as appropriate.

Signed-off-by: Angus Clark <angus.clark@st.com>