stm_spi_fsm: update the JEDEC probe to handle extended READIDs
The previous code was based on 3-byte JEDEC IDs, with a possible 2-byte
extension.  However, devices are now emerging that return 6 or more bytes of
READID data, and the additional bytes are required to differentiate between
variants or generations of device.

This patch refactors the device table and JEDEC probe code to handle arbitrary
length READIDs, with the standard JEDEC definition now becoming a special case.
Functionally, there should be no change in behaviour.  A subsequent patch will
update the table with extended READIDs where applicable.

Note, an equivalent approach for the m25p80.c driver has been adopted in
principle by the MTD community.

Signed-off-by: Angus Clark <angus.clark@st.com>