stm: stxh205: Detect new STxH273 (Palma 2) and STxH237 (Cardiff 2)
Until now all variants of the STxH205 have used the same die, and as
a result it has been impossible to determine at run time which variant
we are running on. However the new STxH273 (Palma 2) and STxH237
(Cardiff 2) variants uses different die, and as a result we can check
the device ID and update the CPU type.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>