stm_spi_fsm: improve busy wait handling
Lee Jones' <lee.jones@linaro.org> work on upstreaming the stm_spi_fsm driver has
revealed a potential race condition in two busy-wait loops: an ill-timed
deschedule could result in the timeout condition being met before any attempt to
test for completion of the operation was made.

In this patch, the race condition is fixed and the fsm_wait_busy() function is
updated to a take a timeout parameter.  This allows us to specify different
timeout delays depending on the operation being performed.  Previously, a fixed,
worst-case delay (corresponding to the Chip Erase operation, ~300s!) was used.
For the moment, we have defined conservative delays for each relevant operation,
which should accommodate all existing devices.  In principle, one could set the
delays according the device probed, but there is probably little to be gained.

Signed-off-by: Angus Clark <angus.clark@st.com>