stm_spi_fsm: s25flxxx: workaround for self-clearing QUAD bit on status writes
A Write Status operation is achieved by a 1-byte WRR command.  However, on
S25FLxxxX devices this has the side effect of clearing the QUAD bit in the
Config register (other bits in the Config register do not seem to be affected).
This behaviour has been confirmed by Spansion engineers and future products have
been redesigned.  The follow extract is from a Spansion FS-S datasheet:

   "WRR Single Register Write

    In some legacy SPI devices, a Write Registers (WRR) command with only one
    data byte would update Status Register 1 and clear some bits in
    Configuration Register 1, including the Quad Mode bit. This could result in
    unintended exit from Quad Mode. The S25FS-S family only updates Status
    Register 1 when a single data byte is provided. The Configuration Register 1
    is not modified in this case"

As a workaround, we first read the Config register, combine with the new Status
data, and then promote to a 2-byte WRR command.

Signed-off-by: Angus Clark <angus.clark@st.com>