stmmac: prevent false carrier sense detection
This patch is to w/a a problem that happens on
some boxes when run at 10Mbps Half duplex mode.

During the transmission the CSR signal is asserted
for some time and the frames aborted because of
carrier sense error. This is reported by MMC HW
counter: txcarrier signal.
This actually is a false carrier so the frames
are good and there is no reason to ask for dropping
them.

This patch so disables the Carrier Sense During Transmission
and this means that the MAC transmitter ignore the CRS signal
during frame transmission in Half-Duplex mode.

Buglink: https://bugzilla.stlinux.com/show_bug.cgi?id=44857

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>